TODO-> Create Material Template

+++ source/gx/IndexBuffer.cpp
@@ -8,7 +8,7 @@ IndexBuffer::IndexBuffer(const GLvoid *data, GLsizei count) : _count(count)
 {
     glGenBuffers(1, &_indexexBufferID);
     glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, _indexexBufferID);
-    glBufferData(GL_ELEMENT_ARRAY_BUFFER, count * sizeof(float), data, GL_STATIC_DRAW);
+    glBufferData(GL_ELEMENT_ARRAY_BUFFER, count * sizeof(unsigned int), data, GL_STATIC_DRAW);
 }
 void IndexBuffer::Bind() const
 {
